Bugfix in the projectile pawn.
When the projectile hit something, it would apply damage to the hit
pawn, which would pass the damage to the controller. This caused some
troubles, where you couldn't be killed by the shooter. Now, the
projectile applies the damage directly to the controller of the hit
pawn, which however solves the problem.